기능 요구사항 - 기능 수정 : 2개의 숫자에 대해 나눗셈이 가능
- util.js parseNumber 로직을 parserFloat 으로 변경

+++ src/js/component/util.js
@@ -27,7 +27,7 @@ const parseNumericalExpression = (rawStr) => {
 }
 
 const parseNumber = (num) => {
-  return parseInt(num)
+  return parseFloat(num)
 }
 
 export { isRightNumericalExpression, parseNumericalExpression }
